requirements.txt
autopep8==2.0.2
blinker==1.6.2
bottle==0.12.25
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
proxy-tools==0.1.0
pycairo==1.23.0
pycodestyle==2.10.0
PyGObject==3.44.1
pywebview==4.1
tomli==2.0.1
watchdog==3.0.0
Werkzeug==2.3.4
